ai-agent/requirements.txt
google-generativeai
firebase-admin
google-cloud-firestore
google-auth
requests
python-dotenv
google-adk